+++ shared/ui/nav-home/nav-home.tsx
@@ -4,9 +4,9 @@ import SearchIcon from "@/shared/assets/svgs/search.svg";
 import { cn } from "@/shared/utils/cn";
 import Link from "next/link";
 
-import * as styles from "./nav-home.css";
-
 import { useState } from "react";
+import HamburgerMenuButton from "../hamburger-menu-button/hamburger-menu-button";
+import * as styles from "./nav-home.css";
 
 const NavHome = () => {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
@@ -23,6 +23,7 @@ const NavHome = () => {
           isMenuOpen ? styles.navHomeStyleOpen : "",
         )}
       >
+        {/* 데스크탑 메뉴 목록 */}
         <div>logo</div>
         <div className={styles.navWrapper}>
           <ul className={styles.navList}>
@@ -45,34 +46,14 @@ const NavHome = () => {
           <button className={styles.searchButtonStyle}>
             <SearchIcon width={"2rem"} height={"2rem"} />
           </button>
-
-          {/* 메뉴버튼 클릭시 햄버거바 -> x버튼으로 rotate */}
-          <button className={styles.menuButtonStyle} onClick={handleMenuClick}>
-            <div className={styles.hamburgerIcon}>
-              <div
-                className={cn(
-                  styles.hamburgerLineTop,
-                  isMenuOpen ? styles.xIconTop : "",
-                )}
-              ></div>
-              <div
-                className={cn(
-                  styles.hamburgerLineMiddle,
-                  isMenuOpen ? styles.xIconMiddle : "",
-                )}
-              ></div>
-              <div
-                className={cn(
-                  styles.hamburgerLineBottom,
-                  isMenuOpen ? styles.xIconBottom : "",
-                )}
-              ></div>
-            </div>
-          </button>
+          <HamburgerMenuButton
+            isMenuOpen={isMenuOpen}
+            onClick={handleMenuClick}
+          />
         </div>
       </nav>
 
-      {/* 800px 이하일때 메뉴 목록 표시 */}
+      {/* 모바일 메뉴 목록 */}
       <ul
         className={cn(
           styles.menuContainer,
